refactor(addmood): organize and split styles
