+++ src/components/Mood/common.styles.jsx
@@ -1,16 +1,3 @@
 import { css } from 'styled-components';
 
-const svgSizes = {
-    default: css`
-        height: ${({ theme }) => theme.iconsSizeL};
-        width: ${({ theme }) => theme.iconsSizeL};
-    `,
-    big: css`
-        height: ${({ theme }) => theme.iconDefaultSize};
-        width: ${({ theme }) => theme.iconDefaultSize};
-    `,
-}
-
-export {
-    svgSizes
-}
+export {}
