+++ src/components/Styles/svgSizes.styles.js
@@ -1,0 +1,16 @@
+import { css } from 'styled-components';
+
+const svgSizes = {
+    default: css`
+        height: ${({ theme }) => theme.iconsSizeL};
+        width: ${({ theme }) => theme.iconsSizeL};
+    `,
+    big: css`
+        height: ${({ theme }) => theme.iconDefaultSize};
+        width: ${({ theme }) => theme.iconDefaultSize};
+    `,
+}
+
+export {
+    svgSizes
+}
